requirements.txt
durable-rules==2.0.20
jupyter==1.0.0
nltk==3.4.5
